gateway/requirements.txt
fastapi
uvicorn
httpx
PyJWT
pydantic[email]
